+++ lokigi/mixins/site_solution_pareto.py
@@ -11,7 +11,7 @@ import seaborn as sns
 import sweetpareto.vis as spv
 import textwrap
 
-from lokigi.utils import _colours_and_styles
+from lokigi.utils import _colours_and_styles, _is_maximise_metric
 
 from lokigi.multiobjective import ParetoMetric
 
@@ -95,9 +95,9 @@ class ParetoMixin:
         or minimisation objective.
         """
         if maxx is None:
-            maxx = x_axis == "proportion_within_coverage_threshold"
+            maxx = _is_maximise_metric(x_axis)
         if maxy is None:
-            maxy = y_axis == "proportion_within_coverage_threshold"
+            maxy = _is_maximise_metric(y_axis)
 
         plot_obj = spv.pareto_plot(
             self.solution_df,
@@ -197,16 +197,8 @@ class ParetoMixin:
 
         for idx, (x_metric, y_metric) in enumerate(metric_pairs):
             ax = axes[idx]
-            current_maxx = (
-                (x_metric == "proportion_within_coverage_threshold")
-                if maxx is None
-                else maxx
-            )
-            current_maxy = (
-                (y_metric == "proportion_within_coverage_threshold")
-                if maxy is None
-                else maxy
-            )
+            current_maxx = _is_maximise_metric(x_metric) if maxx is None else maxx
+            current_maxy = _is_maximise_metric(y_metric) if maxy is None else maxy
             plot_obj = spv.pareto_plot(
                 self.solution_df,
                 x=x_metric,